# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: agent-wire-docling
  version: 0.1.0
  description: |
    HTTP API for the agent-wire-docling prototype. Single source of truth for
    backend ↔ frontend. Frontend client is codegen'd from this file via
    openapi-typescript. All agents code against this; any mid-build gap is
    escalated, not silently edited.
servers:
  - url: http://localhost:8000

# ─────────────────────────────────────────────────────────────────────────────
# Tags
...
paths:
...
  /triage:
    get:
      tags: [batch]
      summary: Read the triage rollup for a batch output directory
      parameters:
        - in: query
          name: output_dir
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Triage doc
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Triage' }
        '404': { $ref: '#/components/responses/NotFound' }
  /triage/retry:
    post:
      tags: [batch]
      summary: Apply user edits to triage.yaml — retry / exclude
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [output_dir]
              properties:
                output_dir: { type: string }
      responses:
        '200':
          description: Retry summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  retried: { type: integer }
                  succeeded: { type: integer }
                  still_failed: { type: integer }
                  excluded: { type: integer }
...
  /batch:
    post:
      tags: [batch]
      summary: Start a batch run using a stratum→pipeline map
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchRequest' }
      responses:
        '200':
          description: Batch job accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
  /jobs/{id}:
    get:
      tags: [batch]
      summary: Batch progress (1Hz poll cadence)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current job state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
  /jobs/{id}/stream:
    get:
      tags: [batch]
      summary: Live progress via Server-Sent Events (optional; polling is the fallback)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: text/event-stream of Job deltas
          content:
            text/event-stream: {}
  /batch/{id}/cancel:
    post:
      tags: [batch]
      summary: Stop claiming new jobs; finish in-flight; return final state
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
  /manifest:
    get:
      tags: [batch]
      summary: Current manifest for the active output folder
      parameters:
        - in: query
          name: output_dir
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Manifest
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Manifest' }
  /taste_sessions:
    post:
      tags: [taste]
      summary: Create a new taste session for a scanned folder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TasteSessionCreate' }
      responses:
        '200':
          description: New session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TasteSession' }
  /taste_sessions/{id}:
    get:
      tags: [taste]
      summary: Rehydrate a session's full state
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Session state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TasteSession' }
    patch:
      tags: [taste]
      summary: Update approvals / pipeline assignments / lock status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TasteSessionPatch' }
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TasteSession' }
  /export:
    post:
      tags: [export]
      summary: Start an export job (manifest-only | manifest+md | full archive)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportRequest' }
      responses:
        '200':
          description: Export job accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
  /exports/{id}:
    get:
      tags: [export]
      summary: Export job progress / result path
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current export state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }

# ─────────────────────────────────────────────────────────────────────────────
# Components
components:
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: string }
...
    PipelineParams:
      type: object
      properties:
        ocr:
          type: object
          properties:
            enabled: { type: boolean, default: true }
            engine: { type: string, enum: [tesseract, rapidocr], default: tesseract }
        vlm:
          type: object
          properties:
            enabled: { type: boolean, default: false }
            model: { type: string, default: 'granite_docling' }
        tables:
          type: object
          properties:
            enabled: { type: boolean, default: true }
        enrichments:
          type: object
          properties:
            formulas: { type: boolean, default: false }
            code: { type: boolean, default: false }
            charts: { type: boolean, default: false }
...
    BatchRequest:
      type: object
      required: [output_dir]
      description: |
        Two dispatch modes:
          - Filemap mode (Level B): provide `root`; optional
            `pipeline_by_content_type` / `pipeline_by_stratum`.
          - Legacy stratum mode: provide `scan_id` + `stratum_pipelines`.
      properties:
        output_dir: { type: string }
        concurrency: { type: integer, default: 2 }
        root: { type: string, nullable: true, description: "Filemap mode: walks .understanding/folder.yaml under this root" }
        pipeline_by_content_type:
          type: object
          nullable: true
          additionalProperties: { $ref: '#/components/schemas/PipelineParams' }
        pipeline_by_stratum:
          type: object
          nullable: true
          additionalProperties: { $ref: '#/components/schemas/PipelineParams' }
        scan_id: { type: string, nullable: true }
        stratum_pipelines:
          type: array
          nullable: true
          description: 'Legacy per-stratum pipeline assignments'
          items:
            type: object
            required: [stratum, pipeline]
            properties:
              stratum: { type: string }
              pipeline: { $ref: '#/components/schemas/PipelineParams' }
...
    Triage:
      type: object
      properties:
        batch_id: { type: string }
        completed_at: { type: string }
        output_dir: { type: string }
        docs_succeeded: { type: integer }
        docs_failed: { type: integer }
        by_reason:
          type: object
          additionalProperties: { type: integer }
        by_content_type:
          type: object
          additionalProperties: { type: integer }
        failures:
          type: array
          items:
            type: object
            additionalProperties: true
    Job:
      type: object
      required: [id, kind, status, progress]
      properties:
        id: { type: string }
        kind: { type: string, enum: [batch, export] }
        status: { type: string, enum: [queued, running, completed, cancelled, failed] }
        progress:
          type: object
          properties:
            docs_total: { type: integer }
            docs_done: { type: integer }
            docs_failed: { type: integer }
            docs_per_sec: { type: number }
            eta_seconds: { type: number, nullable: true }
            per_stratum:
              type: array
              items:
                type: object
                properties:
                  name: { type: string }
                  total: { type: integer }
                  done: { type: integer }
                  failed: { type: integer }
        started_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        error: { type: string, nullable: true }
        result_path: { type: string, nullable: true, description: 'For exports: path to artifact' }

    # ── Per-doc meta & manifest
...
    ManifestEntry:
      type: object
      required: [source_sha256, source_path, source_format, status, stratum, pipeline_hash]
      properties:
        source_sha256: { type: string }
        source_path: { type: string }
        source_format: { type: string }
        status: { type: string, enum: [pending, processing, complete, error, skipped] }
        stratum: { type: string }
        pipeline_hash: { type: string }
        error: { type: string, nullable: true }
        quality_summary:
          type: object
          properties:
            ocr_avg: { type: number, nullable: true }
            warning_count: { type: integer }
            empty_page_count: { type: integer }
        converted_at: { type: string, format: date-time, nullable: true }
    Manifest:
      type: object
      required: [schema_version, folder_root, output_dir, created_at, updated_at, docs]
      properties:
        schema_version: { type: integer, default: 1 }
        folder_root: { type: string }
        output_dir: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        docling_version: { type: string }
        docs:
          type: array
          items: { $ref: '#/components/schemas/ManifestEntry' }

    # ── Taste session
    TasteSessionCreate:
      type: object
      required: [scan_id, output_dir]
      properties:
        scan_id: { type: string }
        output_dir: { type: string }
    DocApproval:
      type: object
      required: [source_sha256, status, reviewed_at]
      properties:
        source_sha256: { type: string }
        status: { type: string, enum: [approved, rejected, skipped, flagged] }
        notes: { type: string, nullable: true }
        reviewed_at: { type: string, format: date-time }
        pipeline_hash: { type: string, description: 'Pipeline under which this doc was reviewed' }
    StratumState:
      type: object
      required: [name, size, pipeline, approvals, locked, status]
      properties:
        name: { type: string }
        size: { type: integer }
        pipeline: { $ref: '#/components/schemas/PipelineParams' }
        approvals:
          type: array
          items: { $ref: '#/components/schemas/DocApproval' }
        locked: { type: boolean }
        status: { type: string, enum: [under_review, converged, non_convergent, exhausted] }
    TasteSession:
      type: object
      required: [id, folder_root, output_dir, scan_id, strata, created_at, updated_at]
      properties:
        id: { type: string }
        folder_root: { type: string }
        output_dir: { type: string }
        scan_id: { type: string }
        strata:
          type: array
          items: { $ref: '#/components/schemas/StratumState' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer, description: 'Monotonic; used for PATCH optimistic locking' }
    TasteSessionPatch:
      type: object
      required: [version]
      properties:
        version: { type: integer, description: 'Current session version for CAS' }
        approval:
          type: object
          description: 'Add or update an approval'
          properties:
            stratum: { type: string }
            approval: { $ref: '#/components/schemas/DocApproval' }
        pipeline_assignment:
          type: object
          description: 'Set a stratum pipeline'
          properties:
            stratum: { type: string }
            pipeline: { $ref: '#/components/schemas/PipelineParams' }
        lock_stratum:
          type: object
          description: 'Lock a stratum (commits pipeline for batch)'
          properties:
            stratum: { type: string }
            locked: { type: boolean }

    # ── Export
    ExportRequest:
      type: object
      required: [output_dir, kind, destination]
      properties:
        output_dir: { type: string }
        kind: { type: string, enum: [manifest_only, manifest_plus_md, full_archive] }
        destination: { type: string, description: 'Absolute file or directory path' }

    # ── DoclingDocument (minimal shape frontend consumes)
...